requirements.txt
fastapi
uvicorn
requests
yt-dlp
python-dotenv

